LayoutTests/media/modern-media-controls/controls-visibility-support/controls-visibility-support-controls-on-expected.txt

Testing the ControlsVisibilitySupport behavior without controls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Media has not begun loading yet
PASS mediaController.controls.controlsBar.visible is false
PASS mediaController.controls.startButton.visible is false

Media has loaded metadata
PASS mediaController.controls.controlsBar.visible is true
PASS mediaController.controls.startButton.visible is true

PASS successfullyParsed is true

TEST COMPLETE
